follow the comments

+++ db/builder.cc
@@ -183,7 +183,7 @@ Status BuildTable(
     std::unique_ptr<BlobFileBuilder> blob_file_builder(
         (mutable_cf_options.enable_blob_files &&
          tboptions.level_at_creation >=
-             static_cast<int>(mutable_cf_options.blob_file_starting_level) &&
+             mutable_cf_options.blob_file_starting_level &&
          blob_file_additions)
             ? new BlobFileBuilder(
                   versions, fs, &ioptions, &mutable_cf_options, &file_options,
